Session cookie implementation (need help)
All changes commented with "cookie" or "question" for easy search. Any help would be much appreciated.

I was trying to add this on to our project, but not sure how it works.
https://firebase.google.com/docs/auth/admin/manage-cookies

So then I found this, npm-cookie, and have imported it, but I'm not sure where I need to establish the cookie. https://www.npmjs.com/package/vue-cookies

I tried setting the cookie @signup, and then on all the pages just to try it out. But Vue only allows one <script> in each Vue component, so I'm having trouble enabling it.