Users structure updated
Able to add new users with the new format (removed username and replaced with fullname), manually modified all the old users so that they now have 'fullname', and are keeping 'username' for now.

New users are able to be added to projects, stages, and tasks, but not channels. This will be solved once you guys copy Mike's userpicker to the chat, since new users are compatible with the dashboard.

Online users now correctly displays user's full name (somehow the old users will still show username instead of fullname).